changed name of connection.ts
