+++ external/js/cothority/src/byzcoin/byzcoin-rpc.ts
@@ -6,6 +6,7 @@ import { Rule } from "../darc";
 import Darc from "../darc/darc";
 import IdentityEd25519 from "../darc/identity-ed25519";
 import IdentityWrapper, { IIdentity } from "../darc/identity-wrapper";
+import { IConnection, LeaderConnection, Roster, RosterWSConnection, WebSocketAdapter } from "../network";
 import { SkipBlock } from "../skipchain/skipblock";
 import SkipchainRPC from "../skipchain/skipchain-rpc";
 import ClientTransaction, { ICounterUpdater } from "./client-transaction";
@@ -29,7 +30,6 @@ import {
     IDVersion,
 } from "./proto/requests";
 import { StreamingRequest, StreamingResponse } from "./proto/stream";
-import {IConnection, LeaderConnection, Roster, RosterWSConnection, WebSocketAdapter} from "../network";
 
 export const currentVersion = 2;
