+++ external/js/cothority/src/calypso/calypso-rpc.ts
@@ -4,7 +4,7 @@ import { Argument, ClientTransaction, InstanceID, Instruction, Proof } from "../
 import ByzCoinRPC from "../byzcoin/byzcoin-rpc";
 import { Signer } from "../darc";
 import { IConnection, Roster, ServerIdentity, WebSocketConnection } from "../network";
-import { RosterWSConnection } from "../network/connection";
+import { RosterWSConnection } from "../network";
 import { registerMessage } from "../protobuf";
 import { DecodeKey, OnChainSecretInstance } from "./calypso-instance";
