+++ external/js/cothority/src/network/websocket.ts
@@ -2,6 +2,7 @@ import { Message } from "protobufjs";
 import { Observable } from "rxjs";
 import Log from "../log";
 import { IConnection } from "./nodes";
+import { Roster } from "./proto";
 import { BrowserWebSocketAdapter, WebSocketAdapter } from "./websocket-adapter";
 
 let factory: (path: string) => WebSocketAdapter = (path: string) => new BrowserWebSocketAdapter(path);
@@ -156,5 +157,22 @@ export class WebSocketConnection implements IConnection {
             });
 
         });
+    }
+}
+
+/**
+ * Single peer connection that reaches only the leader of the roster
+ */
+export class LeaderConnection extends WebSocketConnection {
+    /**
+     * @param roster    The roster to use
+     * @param service   The name of the service
+     */
+    constructor(roster: Roster, service: string) {
+        if (roster.list.length === 0) {
+            throw new Error("Roster should have at least one node");
+        }
+
+        super(roster.list[0].getWebSocketAddress(), service);
     }
 }
